refactor(scheduler): extract trading_session into core/scheduler.py
Moves all trading logic out of main.py into core/scheduler.py so
main.py can become a thin uvicorn entrypoint in the next step. Drops
the runtime.update_trailing_state broadcast call — positions are already
persisted per-pair inside the loop and the in-memory mirror has no
remaining consumer after the Telegram decoupling.
